refactor(WinningNumbers 관련): WinningNumbers 패키지를 dto로 변경
