+++ src/main/java/lotto/domain/WinningRanks.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.dto.WinningNumbers;
+
 import java.util.ArrayList;
 import java.util.List;
